um: handle UM_TIMETRAVEL_RUN only in idle loop, signal success in ACK
The Timetravel socket can be read from the SIGIO handler, which may
catch a RUN-msg that was expected to be received in the wait loop.

It can happen, that the calendar only "created" one interrupt, but an
uncertain(i.e. hard to predict by the calendar) number of SIGIOs is
emitted, where every SIGIO-interrupt produces a GET-msg. The calendar
might already try to send a RUN-msg after the first GET-msg, which is
then processed in the SIGIO handler, waiting for the answer of a
consecutive GET, but time is advanced by the received RUN.  Usually,
this does not imply problems, because the interrupt implies the next
RUN-msg to be at the current time (which was sent by the calendar with
the GET-msg), but there are corner cases, like simultaneous other
interrupts, which may desynchronize the current time in the uml
instance with the expected time from the calendar. Since there is no
real use case for advancing the time in the SIGIO handler with RUN-msgs
(in contrast to UPDATE msgs), we logically only expect to advance the
time in idle in TT-ext mode, therefore, with this patch, we only change
the time in the idle loop.

Additionally, since both do blocking read from the TT-socket, a
deadlock can happen, if a RUN-msg designated for the idle loop is
received in the SIGIO handler. In this situation the calendar expects
the uml instance to run, but actually, it waits for another msg either
in SIGIO handler (e.g. a second interrupt) or waits in poll in the idle
loop as the former msg was handled by the signal handler which gave
back the execution to the main loop which finally ends in the idle loop.

Therefore this patch also allows checking, whether the current RUN-msg
was handled by the idle loop or by the signal/interrupt handlers in the
ACK of the RUN.

With the information in the ACK of the RUN-msg, the calendar knows that
the RUN was answered in a signal handler and can act accordingly.

Signed-off-by: Benjamin <Benjamin.Beichler@uni-rostock.de>