mac80211_hwsim: check skb ptr in _stop operation
We may encounter a race condition when tx-info frames are received and
the interface is stopped. The outcome of skb_queue_empty without the skb
queue lock (which should have been skb_queue_empty_lockless in the first
place) can become outdated when `skb_dequeue` is executed. Consequently,
skb_dequeue can return a NULL, resulting in a NULL-pointer dereference
in ieee80211_free_txskb since there's no check for it.

This patch uses the return of the NULL-pointer to detect an empty queue,
which corrects and simplifies the code. This approach is somewhat more
resource-intensive than using skb_queue_empty_lockless because it
aquires a spinlock for an empty queue. However, this seems reasonable
for the stop operation.

An alternative approach would be to keep the lock for the entire while
loop. However, since ieee80211_free_txskb is responsible for reporting
and other tasks, holding the lock for an extended period could create a
significant critical section and potentially block
hwsim_tx_info_frame_received_nl for longer than acceptable.

Signed-off-by: Benjamin Beichler <benjamin.beichler@uni-rostock.de>